Replace input type number by text for manual error handling
If the user puts in invalid characters, input type number will not complain - just ignore and take 0 as value. This error is not visualized and therefore only recognized when going back to the page. To get an immediate error response, the input type must be text.

Attention: The small arrow buttons the browser adds to input controls are lost.